sanitizeHtml => Use Link, split, map, test functions
sanitizeHtml=> Use the split, map, and test methods.

split(" ") cuts word by word and creates an array
Check if there is a hashtag for each word through the map function
If there is a hastag, use the test function => Put the word in the link as a boolean value.

If there is no hashtag, just wrap the word with <React.Fragment.