Modify InMemoryCache
Only login users are registered as cache in apollo cache.

Solution: Store other user's information in cache & store only ref in root query.

Solution: Add new Apolloclient typePolicies in Apollo.ts.

keyFields: It means the type to be saved. Since id does not exist, it is assigned as username as a unique identifier.