Update Profile.tsx
Create FOLLOW_USER_MUTATION,UNFOLLOW_USER_MUTATION.

If there are multiple conditional statements, make the getButton function to increase the readability of the code.

In the case of isMe, set the Edit Profile button to be visible.

If isFollowing is set, "Unfollowing" and vice versa, make the text "following" visible.

Add PageTitle: "Loading" if loading, make username visible if not loading.