Mutation Follow Unfollow on Profile Page.
Customize your profile page by importing the following, unfollowing mutations.

The other person's profile: edit Profile is not visible, and you can check whether you are following or unfollowing.

Profile of logged-in user: Edit Profile button exists, total Following, and followers are displayed.

It updates the cache so that it can be changed in real time.